backend/requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
